requirements.txt
APScheduler==3.6.3
Django==3.0.3
django-apscheduler==0.3.0
django-cors-headers==3.2.1
djangorestframework==3.11.0
pytz==2019.3
requests==2.23.0
tzlocal==2.0.0
urllib3==1.25.8
pymongo==3.10.1
mongoengine==0.19.1
redis==3.4.1